First iteration on an algorithm library
Implementing tonemapping only for now.
