+++ tsd/src/tsd/CMakeLists.txt
@@ -4,6 +4,7 @@
 # Always present components
 add_subdirectory(app)
 add_subdirectory(core)
+add_subdirectory(algorithms)
 add_subdirectory(io)
 add_subdirectory(animation)
 add_subdirectory(rendering)
