+++ tsd/src/tsd/rendering/CMakeLists.txt
@@ -35,7 +35,7 @@ PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/../..>
 )
 
-project_link_libraries(PUBLIC tsd_scene PRIVATE $<BUILD_INTERFACE:stb_image>)
+project_link_libraries(PUBLIC tsd_scene tsd_algorithms PRIVATE $<BUILD_INTERFACE:stb_image>)
 
 if (TSD_USE_CUDA)
   find_package(thrust REQUIRED HINTS ${CUDAToolkit_LIBRARY_DIR}/cmake)
